Change SRAA institution scope in SamlToken instead of outside of it
Institution configuration options are part of a users context. That is why they should be part of the SamlToken and why they are injected through the constructor. Changing the institution scope as an SRAA is an exception where only an SRAA can modify its current institution (and its institution configuration options).