Use the generic error page for all exceptions
Instead of rendering specific error pages on a case-by-case basis,
handle all exceptions in the stepup-bundle exception controller. This
way, all error pages look the same with only the page title and error
description tailored to the specific error. Logging of exceptions is
also handled by the exception controller.

The following error templates were unused and have been removed:

 * authenticationException.html.twig
 * badCredentials.html.twig
 * noAuthnContext.html.twig
 * unmetLoa.html.twig

The following error templates are removed and the error is handled
using the exception controller:

 * authnFailed.html.twig
 * preconditionNotMet.html.twig
 * blank page in case of 'access denied'

And specific user error messages are shown for the following error
conditions:

 * response is missing a required attribute
 * unknown service provider
 * request not signed
 * request signed with unsupported algorithm
 * request signature validation failed

See:
 - https://www.pivotaltracker.com/story/show/155008515
 - https://www.pivotaltracker.com/story/show/137516239
 - https://www.pivotaltracker.com/story/show/155515714