Fix the AccessDenied exception when no token is set
When a user logs in to the RA with a LoA 3 token and the user does not
have RA rights the access denied error page must be shown.

When authenticating with LoA < 3 does show the access denied error as
it should. Pay attention during testing not to confuse these two.