Remove authorization check from RA
The check if the logged in users institution matches the one of the
token that is to be vetted is removed. This because this check is now
performed in the middleware. And this restrictive check blocks FGA
authorizations, where RAs from institution X should be able to vet
tokens of institution Y.

Error reporting to the end user just became a little bit less verbose.
The ra.form.start_vetting_procedure.different_institution_error flash
message is no longer displayed and the matching notice is no longer
logged. The application falls back on the
ra.form.start_vetting_procedure.unknown_registration_code message
instead.