Expose the FGA options
The MW API now returns the FGA options in the institution-configuration-
options GET endpoint. These values are also made available in the RA.

The InstitutionConfigurationOptionsService also received a
getAvailableInstitutionsFor method, that will show the institutions the
actor is allowed to see/vet/perform actions on.

@pablothedude, please review this method with extra attention. I think
this method should take the actual role of the authenticated identity
into account?