Create and test the AllowedInOtherInstitutionVoter
The voter for now only tests if the logged in user is allowed to
VIEW_AUDITLOG. Other authorizations can be added to this voter if the
context is 'allowed in other institutions'.

The vote method body might need to be revised at that point. But for
now this tests all the appropriate authorization related checks.