Pass the actor id when starting vetting procedure
The actor id is used to test in MW, if the actor is a SRAA. If the user
is SRAA, then it is allowed to vet tokens of any institution. If the
user is not SRAA, then FGA options are taken into account.

For now this additional SRAA context is required when starting the
vetting procedure.

See https://www.pivotaltracker.com/story/show/160283514