Add possibility for RAA to switch institution
The RAA can switch to an institution he/she will be accreditating a
RA candidate on behalf of.

This institution is stored in the SamlToken and persisted between
requests.

The Ra management institution can be found on the SamlToken:
$token->getRaManagementInstitution(). This will be required when
RAAs that can promote users for other RAs is implemented in Stepup-RA.

See: https://www.pivotaltracker.com/story/show/160283517