Add the RAA institution switcher
This switcher is used by RAA users to change between the institutions
they are RAA for.

This switcher is a copy of the SRAA switcher and behaves in a similar
fashion. The difference between them is that the SRAA can switch to all
institutions, and the RAA can only switch to institutions it is actually
RAA for.

The SamlToken was updated and set with a field that stores the 'home'
institution of the RAA. This to ensure the RAA can only switch to his
own institutions.