Set correct institution for authenticating user
In order to authenticate correctly we need to investigate for what
institution(s) the user is listed RA.

If one ra listing is found, that institution is used as the institution
for that RA session. If more than one are retrieved, then the user
should be allowed to select the institution to moderate at a later stage
using the RA switcher. The SHO of the identity is used by default as
the active institution.