Allow all RA roles to change the institution scope
The RA must also be able to switch to the institution it is RA for.
Previously this was only allowed for (S)RAA users via the institution
switcher. Now that the SAML provider also switches to another inst.
outside the SHO of the authenticating user, the change is authorization
was required.