Update authentication in order to find identities
In order to ensure we can find an identity, we no longer need to pass
the institution of that identity. As the RA(A) might be RA for an
institution it does not have a SHO for.

To achieve this goal, the IdentitySearchQuery has been updated, the
Institution parameter has been made optional. This change was applied
in:

 - RA
 - SelfService
 - Middleware-clientbundle
 - Middleware