Custom form rendering on institution configuration
Requirement: The switch button should not float all the way to the
right, either remove it, and submit the form on select change, or align
it more logically. We can also move the form to a more logical place.

This commit updates the way the form is rendered, effectively making
the form a 'one-liner' which floats right, next to the institution
title. When the viewport is to small, the form floats left under the H3