Apply fix from #185 to other paginate calls
Other controller actions use the paginator to paginate their results,
sorting was broken in certain situations.

This was fixed for the token overview in #185, but not fixed for the
other paginator usages.

This should fix the `UnexpectedTypeException` 'PropertyAccessor requires
a graph of objects or arrays to operate on error

See: https://www.pivotaltracker.com/story/show/164729675