Resolve floating issue for role-at-institution
The role at institution select lists are displayed in one row even
though the are two separate elements. The symfony form styling is
overridden in order to render them nicely together. For modern browsers,
flex layout is utilized. For IE10+11, a fallback is used based on the
previous solution that was in place.

See: https://www.pivotaltracker.com/story/show/165719957