Do not error when loa does not existing in vetting procedure
This is a workaround for a case that a LOA does not exist in the
SAML token. The code allows for a nullable loa in the StartVettingProcedureCommand
which will error when checking the required loa level. This changes
checks for existence of the loa and if not will stop the vetting procedure.

See issue: https://github.com/OpenConext/Stepup-RA/issues/246