Upgrade JavaScript dependencies
Many (dev only) security vulnerabilities have been supressed for the past
couple of months. They are now resolved.

Most notable upgrades include a bump to the latest Webpack Encore
The use of Node 14
Update of Less Loader

The last, combined with the update of the fontawesome package  allowed us to
remove the previous fontawesome override action found in the styles.less