Move PHP CS to the new CI/QA location
The previous Ant based run is decomposed in this commit. And moved
to the ci/qa folder. The task is now run using a Composer script
named: phpcs.

PHP cfb is also included (the automatic cs fix thingy). As this is
part of PHPCS, it is also included in this commit