Make qa scripts somewhat more verbose
The yarn audit and composer sec test now have a heading, stating their
script is about to run.

And as an additional bonus. The composer validate action is added to the
phplint script. Having a conflicted lockfile  has often lead to issues
in later stages (post PR). Requiring rework after the work is
'finished'.