Show the recovery token search overview
This is a copy of the second factor overview and shares similar
behavior. So we can search on the relevant RT options like the type,
status. But also search recovery tokens of identities and institutions.

A revocation modal an button was added. Handling of this button is not
yet implemented in this PR.

Translations have been updated and an effort was put into providing a
sane set of English and Dutch translation.