Implement recovery token revocation
This was a simple matter of calling the existing revoke endpoint on
Middleware. Some new boilerplate was created to facilitate this in RA.

A form was added that is triggered by clicking 'continue' in the modal
window. Some filthy JQuery code was copy pasted from the similar SF
revocation solution. This code picks up the recovery token id, actor id
and identity id (the one possessing the RT). And sets the values on the
hidden revocation form.

A command was added that is filled with the form. And it contains the
data required to revoke the token. This command is then posted onto the
Middleware api (command endpoint) after being translated into a correct
MW command in the client bundle.

Translations where updated.

https://www.pivotaltracker.com/story/show/181968091